[Refactor] Extract into SpeechRecognizer.call
